Displayed freqparcoord plot by saving as png; passing in rdata directly instead of csv
TODO: Passing in Rdata directly produces the correct graph, but passing in the equivalent csv produces a different graph. We have to fix the csv to rdata conversion in dsldIsRDataframe(). For now we are passing in rdata directly, but we intend to use the csv and convert it to rdata